+++ src/pharmpy/plugins/nonmem/model.py
@@ -430,11 +430,7 @@ class Model(pharmpy.model.Model):
         error = self._get_error_record()
         if error:
             sub = self.control_stream.get_records('SUBROUTINES')[0]
-            advan = sub.get_option_startswith('ADVAN')
-            trans = sub.get_option_startswith('TRANS')
-            if not trans:
-                trans = 'TRANS1'
-            comp = compartmental_model(self, advan, trans, des)
+            comp = compartmental_model(self, sub.advan, sub.trans, des)
             if comp is not None:
                 cm, link = comp
                 statements += [cm, link]